+++ src/main/java/de/ukbonn/mwtek/dashboard/configuration/FhirServerRestConfiguration.java
@@ -23,7 +23,7 @@ import org.springframework.context.annotation.Configuration;
 
 /**
  * This class is used to read out the individual configuration of the FHIR servers of the data
- * providers
+ * providers.
  *
  * @author <a href="mailto:david.meyers@ukbonn.de">David Meyers</a>
  */