+++ src/main/java/de/ukbonn/mwtek/dashboard/interfaces/SearchService.java
@@ -25,7 +25,7 @@ import org.hl7.fhir.r4.model.Bundle.BundleLinkComponent;
 
 /**
  * Service for providing the {@link FhirSearchConfiguration} (e.g. which Loinc codes are to be used)
- * from application.yaml
+ * from <code>application.yaml</code>.
  *
  * @author <a href="mailto:david.meyers@ukbonn.de">David Meyers</a>
  */