+++ src/main/java/de/ukbonn/mwtek/dashboard/services/AcuwaveSearchService.java
@@ -33,7 +33,7 @@ import org.springframework.web.client.RestTemplate;
 
 /**
  * Service for providing the {@link AcuwaveServerRestConfiguration} (e.g. which Loinc codes are to
- * be used) from application.yaml
+ * be used) from <code>application.yaml</code>.
  *
  * @author <a href="mailto:david.meyers@ukbonn.de">David Meyers</a>
  */